requirements.txt
PyYAML==6.0
tensorboard==2.13.0
timm==0.6.11
transformers==4.28.1
tqdm==4.65.0
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.2
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
xgboost==2.0.3
rasterio==1.3.6
shapely==2.0.1
geopandas==0.12.2
pyproj==3.5.0